# laravel/source/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/salepromotion": {
            "get": {
                "tags": [
                    "SalePromotion"
                ],
                "summary": "Get All SalePromotion",
                "operationId": "9ce58c722dd8c4f225c63cac12620c71",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "SalePromotion"
                ],
                "summary": "Add SalePromotion",
                "operationId": "b3a78726ba830f3579ff0eb67fd7e762",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "discount": {
                                        "type": "integer"
                                    },
                                    "timeStart": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "timeEnd": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "visible": {
                                        "type": "integer"
                                    },
                                    "mobileBanner": {
                                        "type": "string"
                                    },
                                    "pcBanner": {
                                        "type": "string"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/salepromotion/{id}": {
            "get": {
                "tags": [
                    "SalePromotion"
                ],
                "summary": "Get SalePromotion",
                "operationId": "5162baa99635082cdcd3d6de52d8e712",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "SalePromotion"
                ],
                "summary": "Update SalePromotion",
                "operationId": "0e2549849744b8ce352ccd0e47bb02ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "discount": {
                                        "type": "integer"
                                    },
                                    "timeStart": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "timeEnd": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "visible": {
                                        "type": "integer"
                                    },
                                    "mobileBanner": {
                                        "type": "string"
                                    },
                                    "pcBanner": {
                                        "type": "string"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SalePromotion"
                ],
                "summary": "Delete SalePromotion",
                "operationId": "e5bd0f0e726ed17ada4ec4aaa746e6d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/size": {
            "get": {
                "tags": [
                    "Size"
                ],
                "summary": "Get All Size",
                "operationId": "b16f635ab6feb58eddbb871f3872e6a0",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Size"
                ],
                "summary": "Add Size",
                "operationId": "21c1192edd6b089d98b149119a079eea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "variantId": {
                                        "type": "integer"
                                    },
                                    "size": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/size/{id}": {
            "get": {
                "tags": [
                    "Size"
                ],
                "summary": "Get Size",
                "operationId": "5754368bc14669a6b09abfb6f71d856a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Size"
                ],
                "summary": "Update Size",
                "operationId": "96ae4a38b4d2e4fc2dcd154dda0242df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "variantId": {
                                        "type": "integer"
                                    },
                                    "size": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Size"
                ],
                "summary": "Delete Size",
                "operationId": "9203b09e4f108c3bbff22c2f6424730b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}